// libs/features/benefits/src/lib/shared/models/benefit-response.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText } from '@gosi-ui/core';

export class BenefitResponse {
  benefitRequestId: number;
  referenceNo: number;
  message: BilingualText;
  hideBenefitDetails?: boolean;
}